Player golden moments
A query that works out the highest Star Quality game they've had in the last month for a specific player.

It then works out all the players that they beat, adds some basic information about the game, and spits it all out.

To be combined with some aggregate information like
"Average Star Quality" and "games played this month", "achievements earned"