Corrected missing participation
Issue was caused when games were created.
When finding an existing game, its uuid was returned.
When creating a new game, the uuid was not returned.

This meant that when adding participation for the 1st player, no uuid was supplied, so the participation wasn't attributed to any game - for the first player to be involved in a game.

Because C'tri and Rayth are typically the first players on the "interesting" list, they suffered most often.