new View - playerHomes, new table ArenaRanksLookup
playerHomes figures out on a month-to-month basis which arena the plahyer has played most in. That is their "home" for the month. 
To be used for coming up with suitable titles for traveling players in the big 5

ArenaRanks will be used to look up the player skill number with the text in a way that's arena specific. This will handle foreign arenas with localised ranks, and arenas with ranks outside the default 6.

Intended population - only to occur on targeting individuals. Onboarding experience will need to select distinct examples and draw them in.